requirements.txt
certifi @ file:///croot/certifi_1665076670883/work/certifi
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
docopt==0.6.2
fonttools==4.38.0
github-clone==1.2.0
idna==3.4
imageio==2.22.2
kiwisolver==1.4.4
liblinear-official==2.45.0
libsvm-official==3.30.0
matplotlib==3.6.1
networkx==2.8.7
numpy==1.23.4
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
requests==2.28.1
SciencePlots==1.0.7
scikit-image==0.19.3
scipy==1.9.3
six==1.16.0
tifffile==2022.10.10
torch @ file:///mnt/d/wsl_files/pkgs/torch-1.12.1-cp310-cp310-manylinux1_x86_64.whl
torchvision @ file:///mnt/d/wsl_files/pkgs/torchvision-0.13.1-cp310-cp310-manylinux1_x86_64.whl
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
